feat(Views/v2) add the PartialTestCase
In the context of the work done for Views v2 we've added the
`HtmlPartialTestCase` hard-wiring, into it, the use of the snapshots
library.

This PR adds a second implementation of the partial test case (to avoid
back-compatibility issues) that will **not** assume the use of a
specific snapshot library to allow projects using it to use what they
want/need.