fix(With_Post_Remapping) restore event meta
When mocking an Event, we use the `tribe_get_event` function to return a
post object for the mocked information. If a call to `add_post_meta` or
`update_post_meta` happens in the context of hydrating that event
properties, then the mock meta in the cache will be wiped.
This commit fixes this, by making sure the original mock meta will be
restored after getting the object.